Public Functions

Finch.h.FINCH_SCOPE — Function

FINCH_SCOPE([stmt])

Execute the statement stmt in a new finch scope. All finch objects allocated within this scope will be freed when the scope is closed, unless passed to finch_escape to pass them to the parent scope. The user must not use return or break to leave stmt.

Finch.h.finch_escape — Function

jl_value_t* finch_escape(jl_value_t* var)

Removes var from the current scope and registers it with the parent scope. This meansvar will not be freed when the current scope is closed.

Finch.h.finch_eval — Function

jl_value_t* finch_eval(const char* proc)

Evaluate the Julia code represented by the string proc at global scope in the Main module.

Finch.h.finch_consume_vector — Function

jl_value_t* finch_consume_vector(jl_datatype_t* type, void* ptr, int len);

Create a Julia array with elements of datatype type from the pointer ptr. The array will be of length len, no copying will be performed, and Finch may call free(ptr).

Finch.h.finch_free — Function

void finch_free(jl_value_t* var)

Unregister the Finch-tracked object var within the current scope to allow the garbage collector to free memory. This method should be avoided in favor of using FINCH_SCOPE to limit the lifetime of objects.

Finch.h.finch_mirror_vector — Function

jl_value_t* finch_mirror_vector(jl_datatype_t* type, void* ptr, int len);

Create a Julia array with elements of datatype type from the pointer ptr. The array will be of length len, no copying will be performed, and Finch may not call free(ptr).

Finch.h.finch_initialize — Function

void finch_initialize()

Initialize Finch. Should be called only once before any other finch calls, from the executable.

Finch.h.finch_root — Function

jl_value_t* finch_root(jl_value_t* var)

Register the Julia-allocated object var with Finch on the current scope to avoid garbage collecting it.

Finch.h.finch_exec — Function

jl_value_t* finch_exec(const char* proc, jl_value_t* args...)

Evaluate the Julia code represented by the string proc at local scope in the Main module. proc can optionally contain format specifiers to interpolate julia arguments. Format specifiers should be either %s for a julia input or %% for a literal % character. For example,

    finch_exec("%s + %s", x, y)

should evaluate to x + y

finch_exec caches inputs by their string to avoid repeated compilation.

Finch.h.finch_T — Function

void finch_[T](S x);

Create a Julia object of type T from corresponding C object x of type S.

Finch.h.finch_call — Function

jl_value_t* finch_call(jl_value_t* f, jl_value_t* args...)

Call the Julia function f on the arguments args and return the result. This is a macro that counts the number of arguments.

Finch.h.finch_finalize — Function

void finch_finalize()

Finalize Finch. Should be called at the end of the program to allow Finch to cleanup.