Finch Tensor File Input/Output

Finch supports many input/output tensor file formats.

Finch Format (.fbr)

Finch's custom binary file format for fibers is best suited to users who plan to exclusively use Finch in Julia (perhaps across different platforms). This format straightforwardly maps the fields of Finch fiber formats to arrays in data containers (currently Finch only supports HDF5, but if you file an issue someone might add Numpy NPZ). Arrays are stored 1-indexed as they would be in memory.

Finch.fbrwrite — Function

fbrwrite(filename, tns)

Write the Finch fiber using Finch's custom binary HDF5 file format.

HDF5 must be loaded for this function to be available

Finch.fbrread — Function

fbrread(filename)

Read the Finch fiber using Finch's custom binary HDF5 file format.

HDF5 must be loaded for this function to be available

Binsparse Format (.fbr)

Finch supports the most recent revision of the Binsparse binary sparse tensor format, as well as the proposed v2.0 tensor extension. This is a good option for those who want an efficient way to transfer sparse tensors between supporting libraries and languages. The Binsparse format represents the tensor format as a JSON string in the underlying data container (currently Finch only supports HDF5, but if you file an issue someone might add Numpy NPZ). Binsparse arrays are stored 0-indexed.

Finch.bswrite — Function

bswrite(filename, tns)

Write the Finch fiber to a file using Binsparse HDF5 file format.

HDF5 must be loaded for this function to be available

Warning

The Binsparse spec is under development. Additionally, this function may not be fully conformant. Please file bug reports if you see anything amiss.

Finch.bsread — Function

bsread(filename)

Read the Binsparse HDF5 file into a Finch tensor.

HDF5 must be loaded for this function to be available

Warning

The Binsparse spec is under development. Additionally, this function may not be fully conformant. Please file bug reports if you see anything amiss.

TensorMarket (.mtx, .ttx)

Finch supports the MatrixMarket and TensorMarket formats, which prioritize readability and archiveability, storing matrices and tensors in plaintext.

Finch.fttwrite — Function

fttwrite(filename, tns)

Write a sparse Finch fiber to a TensorMarket file.

TensorMarket must be loaded for this function to be available.

See also: ttwrite

Finch.fttread — Function

fttread(filename, infoonly=false, retcoord=false)

Read the TensorMarket file into a Finch fiber. The fiber will be dense or COO depending on the format of the file.

TensorMarket must be loaded for this function to be available.

See also: ttread