Finch Compilation Pipeline

Finch Notation

Finch.FinchNotation.FinchNode — Type

FinchNode

A Finch IR node. Finch uses a variant of Concrete Index Notation as an intermediate representation.

The FinchNode struct represents many different Finch IR nodes. The nodes are differentiated by a FinchNotation.FinchNodeKind enum.

Finch.FinchNotation.literal — Constant

literal(val)

Finch AST expression for the literal value val.

Finch.FinchNotation.index — Constant

index(name)

Finch AST expression for an index named name. Each index must be quantified by a corresponding loop which iterates over all values of the index.

Finch.FinchNotation.variable — Constant

variable(name)

Finch AST expression for a variable named name. The variable can be looked up in the context.

Finch.FinchNotation.virtual — Constant

virtual(val)

Finch AST expression for an object val which has special meaning to the compiler. This type allows users to substitute their own ASTs, etc. into Finch expressions.

Finch.FinchNotation.value — Constant

value(val, type)

Finch AST expression for host code val expected to evaluate to a value of type type.

Finch.FinchNotation.loop — Constant

loop(idx, ext, body)

Finch AST statement that runs body for each value of idx in ext. Tensors in body must have ranges that agree with ext.

Finch.FinchNotation.assign — Constant

assign(lhs, op, rhs)

Finch AST statement that updates the value of lhs to op(lhs, rhs). The tensors of lhs are returned. Overwriting is accomplished with the function right(lhs, rhs) = rhs.

Finch.FinchNotation.call — Constant

call(op, args...)

Finch AST expression for the result of calling the function op on args....

Finch.FinchNotation.cached — Constant

cached(val, ref)

Finch AST expression val, equivalent to the quoted expression ref

Finch.FinchNotation.reader — Constant

reader()

Finch AST expression for an access mode that is read-only.

Finch.FinchNotation.updater — Constant

updater(mode)

Finch AST expression for an access mode that may modify the tensor. The mode field specifies whether this access returns this tensor (initializing and finalizing) or modifies it in place.

Finch.FinchNotation.create — Constant

create()

Finch AST expression for an "allocating" update. This access will initialize and freeze the tensor, and we can be sure that any values the tensor held before have been forgotten.

Finch.FinchNotation.modify — Constant

modify()

Finch AST expression for an "in place" update. The access will not initialize or freeze the tensor, but can modify it's existing values.

Finch.FinchNotation.declare — Constant

declare(tns, init)

Finch AST statement that declares tns with an initial value init in the current scope.

Finch.FinchNotation.freeze — Constant

freeze(tns)

Finch AST statement that freezes tns in the current scope.

Finch.FinchNotation.thaw — Constant

thaw(tns)

Finch AST statement that thaws tns in the current scope.

Finch.FinchNotation.forget — Constant

forget(tns)

Finch AST statement that marks the end of the lifetime of tns, at least until it is declared again.

Finch.FinchNotation.access — Constant

access(tns, mode, idx...)

Finch AST expression representing the value of tensor tns at the indices idx.... The mode differentiates between reads or updates and whether the access is in-place.

Finch.FinchNotation.sequence — Constant

sequence(bodies...)

Finch AST statement that executes each of it's arguments in turn. If the body is not a sequence, replaces accesses to read-only tensors in the body with getreader and accesses to update-only tensors in the body with getupdater.

Finch.FinchNotation.protocol — Constant

protocol(idx, mode)

Finch AST expression marking an indexing expression idx with the protocol mode. These usually reside at the toplevel of an indexing expression to an access.

Finch.FinchNotation.sieve — Constant

sieve(cond, body)

Finch AST statement that only executes body if cond is true.

Finch.FinchNotation.finch_leaf — Function

finch_leaf(x)

Return a terminal finch node wrapper around x. A convenience function to determine whether x should be understood by default as a literal, value, or virtual.

Program Instances

Finch relies heavily on Julia's metaprogramming capabilities ( macros and generated functions in particular) to produce code. To review briefly, a macro allows us to inspect the syntax of it's arguments and generate replacement syntax. A generated function allows us to inspect the type of the function arguments and produce code for a function body.

In normal Finch usage, we might call Finch as follows:

julia> C = @fiber(sl(e(0)));

julia> A = @fiber(sl(e(0)), [0, 2, 0, 0, 3]);

julia> B = @fiber(d(e(0)), [11, 12, 13, 14, 15]);

julia> @finch (C .= 0; @loop i C[i] = A[i] * B[i]);

julia> C
SparseList (0) [1:5]
├─[2]: 24
├─[5]: 45

The @macroexpand macro allows us to see the result of applying a macro. Let's examine what happens when we use the @finch macro (we've stripped line numbers from the result to clean it up):

julia> (@macroexpand @finch (C .= 0; @loop i C[i] = A[i] * B[i])) |> Finch.striplines |> Finch.regensym
quote
    _res_1 = (Finch.execute)((Finch.FinchNotation.sequence_instance)((Finch.FinchNotation.declare_instance)((Finch.FinchNotation.variable_instance)(:C, (Finch.FinchNotation.finch_leaf_instance)(C)), literal_instance(0)), begin
                    let i = index_instance(i)
                        (Finch.FinchNotation.loop_instance)(i, index_instance(:), (Finch.FinchNotation.assign_instance)((Finch.FinchNotation.access_instance)((Finch.FinchNotation.variable_instance)(:C, (Finch.FinchNotation.finch_leaf_instance)(C)), (Finch.FinchNotation.updater_instance)((Finch.FinchNotation.create_instance)()), (Finch.FinchNotation.variable_instance)(:i, (Finch.FinchNotation.finch_leaf_instance)(i))), (Finch.FinchNotation.literal_instance)((Finch.FinchNotation.initwrite)((Finch.default)(C))), (Finch.FinchNotation.call_instance)((Finch.FinchNotation.variable_instance)(:*, (Finch.FinchNotation.finch_leaf_instance)(*)), (Finch.FinchNotation.access_instance)((Finch.FinchNotation.variable_instance)(:A, (Finch.FinchNotation.finch_leaf_instance)(A)), (Finch.FinchNotation.reader_instance)(), (Finch.FinchNotation.variable_instance)(:i, (Finch.FinchNotation.finch_leaf_instance)(i))), (Finch.FinchNotation.access_instance)((Finch.FinchNotation.variable_instance)(:B, (Finch.FinchNotation.finch_leaf_instance)(B)), (Finch.FinchNotation.reader_instance)(), (Finch.FinchNotation.variable_instance)(:i, (Finch.FinchNotation.finch_leaf_instance)(i))))))
                    end
                end))
    begin
        C = Finch.get(_res_1, :C, C)
    end
    begin
        _res_1
    end
end

In the above output, @finch creates an AST of program instances, then calls Finch.execute on it. A program instance is a struct that contains the program to be executed along with its arguments. Although we can use the above constructors (e.g. loop_instance) to make our own program instance, it is most convenient to use the unexported macro Finch.finch_program_instance:

julia> using Finch: @finch_program_instance

julia> prgm = Finch.@finch_program_instance (C .= 0; @loop i C[i] = A[i] * B[i])
sequence_instance(declare_instance(variable_instance(:C, C), literal_instance(0)), loop_instance(index_instance(i), index_instance(:), assign_instance(access_instance(variable_instance(:C, C), updater_instance(create_instance()), index_instance(i)), literal_instance(Finch.FinchNotation.InitWriter{0}()), call_instance(variable_instance(:*, *), access_instance(variable_instance(:A, A), reader_instance(), index_instance(i)), access_instance(variable_instance(:B, B), reader_instance(), index_instance(i))))))

As we can see, our program instance contains not only the AST to be executed, but also the data to execute the program with. The type of the program instance contains only the program portion; there may be many program instances with different inputs, but the same program type. We can run our program using Finch.execute, which returns a NamedTuple of outputs.

julia> typeof(prgm)
Finch.FinchNotation.SequenceInstance{Tuple{Finch.FinchNotation.DeclareInstance{Finch.FinchNotation.VariableInstance{:C, Fiber{SparseListLevel{Int64, Int64, ElementLevel{0, Int64}}}}, Finch.FinchNotation.LiteralInstance{0}}, Finch.FinchNotation.LoopInstance{Finch.FinchNotation.IndexInstance{:i}, Finch.FinchNotation.IndexInstance{:(:)}, Finch.FinchNotation.AssignInstance{Finch.FinchNotation.AccessInstance{Finch.FinchNotation.VariableInstance{:C, Fiber{SparseListLevel{Int64, Int64, ElementLevel{0, Int64}}}}, Finch.FinchNotation.UpdaterInstance{Finch.FinchNotation.CreateInstance}, Tuple{Finch.FinchNotation.IndexInstance{:i}}}, Finch.FinchNotation.LiteralInstance{Finch.FinchNotation.InitWriter{0}()}, Finch.FinchNotation.CallInstance{Finch.FinchNotation.VariableInstance{:*, Finch.FinchNotation.LiteralInstance{*}}, Tuple{Finch.FinchNotation.AccessInstance{Finch.FinchNotation.VariableInstance{:A, Fiber{SparseListLevel{Int64, Int64, ElementLevel{0, Int64}}}}, Finch.FinchNotation.ReaderInstance, Tuple{Finch.FinchNotation.IndexInstance{:i}}}, Finch.FinchNotation.AccessInstance{Finch.FinchNotation.VariableInstance{:B, Fiber{DenseLevel{Int64, ElementLevel{0, Int64}}}}, Finch.FinchNotation.ReaderInstance, Tuple{Finch.FinchNotation.IndexInstance{:i}}}}}}}}}

julia> C = Finch.execute(prgm).C
SparseList (0) [1:5]
├─[2]: 24
├─[5]: 45

This functionality is sufficient for building finch kernels programatically. For example, if we wish to define a function pointwise_sum() that takes the pointwise sum of a variable number of vector inputs, we might implement it as follows:

julia> function pointwise_sum(As...)
           B = @fiber(d(e(0)))
           isempty(As) && return B
           i = Finch.FinchNotation.index_instance(:i)
           A_vars = [Finch.FinchNotation.variable_instance(Symbol(:A, n), As[n]) for n in 1:length(As)]
           #create a list of variable instances with different names to hold the input tensors
           ex = @finch_program_instance 0
           for A_var in A_vars
               ex = @finch_program_instance $A_var[i] + $ex
           end
           prgm = @finch_program_instance (B .= 0; @loop i B[i] = $ex)
           return Finch.execute(prgm).B
       end
pointwise_sum (generic function with 1 method)

julia> pointwise_sum([1, 2], [3, 4])
Dense [1:2]
├─[1]: 4
├─[2]: 6

Virtualization

TODO more on the way...

Tensor Life Cycle

Every virtual tensor must be in one of two modes: read-only mode or update-only mode. The following functions may be called on virtual tensors throughout their life cycle.

Finch.declare! — Function

declare!(tns, ctx, init)

Declare the read-only virtual tensor tns in the context ctx with a starting value of init and return it. Afterwards the tensor is update-only.

Finch.get_reader — Function

get_reader(tns, ctx, protos...)

Return an object (usually a looplet nest) capable of reading the read-only virtual tensor tns. As soon as a read-only tensor enters scope, each subsequent read access will be initialized with a separate call to get_reader. protos is the list of protocols in each case.

Finch.get_updater — Function

get_updater(tns, ctx, protos...)

Return an object (usually a looplet nest) capable of updating the update-only virtual tensor tns. As soon as an update only tensor enters scope, each subsequent update access will be initialized with a separate call to get_updater. protos is the list of protocols in each case.

Finch.freeze! — Function

freeze!(tns, ctx)

Freeze the update-only virtual tensor tns in the context ctx and return it. Afterwards, the tensor is read-only.

Finch.trim! — Function

trim!(tns, ctx)

Before returning a tensor from the finch program, trim any excess overallocated memory.

Finch.thaw! — Function

thaw!(tns, ctx)

Thaw the read-only virtual tensor tns in the context ctx and return it. Afterwards, the tensor is update-only.

Fiber internals

Fiber levels are implemented using the following methods:

Finch.default — Function

default(arr)

Return the initializer for arr. For SparseArrays, this is 0. Often, the default value becomes the fill or background value of a tensor.

Finch.declare_level! — Function

declare_level!(lvl, ctx, pos, init)

Initialize and thaw all fibers within lvl, assuming positions 1:pos were previously assembled and frozen. The resulting level has no assembled positions.

Finch.assemble_level! — Function

assemble_level!(lvl, ctx, pos, new_pos)

Assemble and positions pos+1:new_pos in lvl, assuming positions 1:pos were previously assembled.

Finch.reassemble_level! — Function

reassemble_level!(lvl, ctx, pos_start, pos_end)

Set the previously assempled positions from pos_start to pos_end to level_default(lvl).

Finch.freeze_level! — Function

freeze_level!(lvl, ctx, pos)

Freeze all fibers in lvl. Positions 1:pos need freezing.

Finch.level_ndims — Function

level_ndims(::Type{Lvl})

The result of level_ndims(Lvl) defines ndims for all subfibers in a level of type Lvl.

Finch.level_size — Function

level_size(lvl)

The result of level_size(lvl) defines the size of all subfibers in the level lvl.

Finch.level_axes — Function

level_axes(lvl)

The result of level_axes(lvl) defines the axes of all subfibers in the level lvl.

Finch.level_eltype — Function

level_eltype(::Type{Lvl})

The result of level_eltype(Lvl) defines eltype for all subfibers in a level of type Lvl.

Finch.level_default — Function

level_default(::Type{Lvl})

The result of level_default(Lvl) defines default for all subfibers in a level of type Lvl.

Debugging Functionality

It's easy to ask Finch to advance a few steps in its compiler pipeline. The basic functionality is documented via the following bit of code:

using Finch
using Finch: @finch_program_instance, begin_debug, step_code, iscompiled, end_debug

y = @fiber d(e(0.0))
A = @fiber d(sl(e(0.0)))
x = @fiber sl(e(0.0))

code = Finch.@finch_program_instance begin
   @loop j i y[i] += A[i, j] * x[j]
end

debug = begin_debug(code)

while true
    global debug = step_code(debug) # Runs one step of compilation
    if iscompiled(debug.code) # Checks if we are done compiling.
        break
    end
end
        
ret = end_debug(debug) # extracts code from debugging context.
# Prints compiled code

The function begin_debug(code; algebra) takes a finch_program_instance plus an optional algebra and creates a debugging context for it, called a PartialCode. The function step_code(debug; steps, sdisplay) takes a debugging context and advances some number of steps, displaying the results automatically if sdisplay. Finally, iscompiled checks if the code in a debug context is completely compiled and end_debug extracts the code, throwing an error if the code is not completely compiled.

Finch.begin_debug — Function

begin_debug(code; algebra = DefaultAlgebra(), sdisplay=false)

Takes a Finch Program and stages it within a DebugContext, defined within a particualr algebra.

Finch.step_code — Function

step_code(code::PartialCode; step=1, sdisplay=false)

Advance the compiler on code for step, displaying the code at the end if sdisplay.

Finch.iscompiled — Function

iscompiled(code:: Expr)

Checks if Julia AST has any Resumables in it.

Finch.end_debug — Function

end_debug(code:: PartialCode)

Returns a finished Julia AST from a PartialCode` if it has no Resumables in it. Throws an error otherwise.

Finch.PartialCode — Type

PartialCode(lastControl :: AbstractLoweringControl, code)

Essentially a debugging context that holds the code that we are working on and the AbstractLoweringControl.

Partially compiled code will be displayed almost like fully compiled code but with @finch nodes that are numbered according to which will be compiled first. They also display where they will renter the compilation pipeline. An early step in the above program might look like:

quote
    y_lvl = ex.body.body.lhs.tns.tns.lvl
    y_lvl_2 = y_lvl.lvl
    A_lvl = (ex.body.body.rhs.args[1]).tns.tns.lvl
    A_lvl_2 = A_lvl.lvl
    A_lvl_3 = A_lvl_2.lvl
    x_lvl = (ex.body.body.rhs.args[2]).tns.tns.lvl
    x_lvl_2 = x_lvl.lvl
    y_lvl.shape == A_lvl_2.shape || throw(DimensionMismatch("mismatched dimension limits ($(y_lvl.shape) != $(@finch((Number = 0, Which = ("lower.jl", 174))A_lvl_2.shape::Int64)))"))
    @finch((Number = 1, Which = ("lower.jl", 174))A_lvl.shape::Int64) == @finch((Number = 2, Which = ("lower.jl", 174))x_lvl.shape::Int64) || throw(DimensionMismatch("mismatched dimension limits ($(@finch((Number = 3, Which = ("lower.jl", 174))A_lvl.shape::Int64)) != $(@finch((Number = 4, Which = ("lower.jl", 174))x_lvl.shape::Int64)))"))
    @finch begin(Number = 5, Which = ("lower.jl", 174))
      begin
        @thaw(y)
        @∀ j = virtual(Finch.Extent) i = virtual(Finch.Extent)  (
          y[i] <<+>>= *(x[j], A[i, j])        )
        @freeze(y)
      end
    end
    qos = @finch((Number = 6, Which = ("lower.jl", 174))1) * @finch((Number = 7, Which = ("lower.jl", 174))y_lvl.shape::Int64)
    resize!(y_lvl_2.val, @finch((Number = 8, Which = ("lower.jl", 174))qos))
    (y = @finch((Number = 9, Which = ("fibers.jl", 27))VirtualFiber(d(e(0.0)))),)
end

Dangers

This feature is experimental and could easily break. In particular, this feature assumes that the Finch compiler never produces code that is needed to produce the next bit of code without putting the required analysis in the program that is being compiled. For example, we cannot currently pause

code = ctx(A1)
info = analysis(code)
code1 = ctx(A2, info)
begin
$code1
$code2
end

because analysis on the first bit of code will give the wrong results if we don't finish it, but it would be okay to do something like this if we placed the analysis in the resulting code, as part of the runtime.

If in the future, Finch needs to do this, this feature will break. However, there is an internal mechanism to recover. The AbstractLoweringControl type is supposed to manage when code is allowed to be paused via the should_pause function. Modifying this function on StepOnlyControl <: AbstractLoweringControl or creating a new control is a potential route to ensure incremental compilation does not occur when it is impossible i.e when dependencies in the compiler mean code cannot be partially compiled.

Finch.Resumable — Type

Resumable(ctx, root, style, meta)

Struct to hold a paused compilation. Holds the compiler state in ctx, a FinchNode in root, and the compiler style in style. If the resumable is in an expression context. The meta field is a Dict of metadata about the resumable.

Finch.should_resume — Function

should_resume(c :: AbstractLoweringControl, ctx, node, style, meta)

Determines if a control wants to resume lowering a Resumable. Should return true or false. Defaults to true.

Finch.should_pause — Function

should_pause(c :: AbstractLoweringControl, ctx, node, style)

Determines if a control wants to pause on a particular lowering of a particular FinchNode. Should return true or false. Defaults to false.

Finch.evolve_control — Function

evolve_control(c :: AbstractLoweringControl, ctx, node, style)

In the event that control does not pause on a Finch node, we evolve the control, resulting in a new AbstractLoweringControl.

Finch.init_meta — Function

init_meta(c :: AbstractLoweringControl, ctx, node, style)

Returns recomputed metadata for a resumable in the event that it is not resumed. Defaults to the empty dict.

Advanced Dangers: Non-Serial Compilation

Furthermore, the Finch compiler is inhernetly serial: statements in a sequence rely on information found via compiling earlier statements. Thus, although this feature exports functions that can reorder the compilation, we do not expect these to work consistently and we leave them basically undocumented.

Finch.step_all_code — Function

step_all_code(code::PartialCode; step=1, sdisplay=false)

Experimental feature: Do not use explictly.

Finch.repeat_step_code — Function

repeat_step_code(code::PartialCode; control=nothing, sdisplay=false)

Experimental feature: Do not use explictly.

Finch.step_some_code — Function

step_some_code(code::PartialCode; step=1, resumeLocations=nothing, resumeStyles=nothing, resumeFilter=nothing, sdisplay=false)

Experimental feature: Do not use explictly.