Public Functions

Finch.fiber — Function

fiber(arr, default = zero(eltype(arr)))

Copies an array-like object arr into a corresponding, similar Fiber datastructure. default is the default value to use for initialization and sparse compression.

See also: fiber!

Examples

julia> println(summary(fiber(sparse([1 0; 0 1]))))
2×2 @fiber(d(sl(e(0))))

julia> println(summary(fiber(ones(3, 2, 4))))
3×2×4 @fiber(d(d(d(e(0.0)))))

Finch.fiber! — Function

fiber!(arr, default = zero(eltype(arr)))

Like fiber, copies an array-like object arr into a corresponding, similar Fiber datastructure. However, fiber! reuses memory whenever possible, meaning arr may be rendered unusable.

Missing docstring.

Missing docstring for sparse. Check Documenter's build log for details.

Missing docstring.

Missing docstring for sparse!. Check Documenter's build log for details.