Public Functions

Finch.CloseScope — Type

CloseScope(ctx)

A transformation to freeze output virtual tensors before they leave scope and are returned to the caller.

See also: freeze!

Finch.DenseData — Type

DenseData(lvl)

Represents a tensor A where each A[i, :, ..., :] is represented by lvl.

Finch.ElementData — Type

ElementData(eltype, default)

Represents a scalar element of type eltype and default default.

Finch.Fiber — Type

Fiber(lvl)

Fiber represents the root of a level-tree tensor. To easily construct a valid fiber, use @fiber or fiber. Users should avoid calling this constructor directly.

In particular, Fiber represents the tensor at position 1 of lvl. The constructor Fiber(lvl) wraps a level assuming it is already in a valid state. The constructor Fiber!(lvl) first initializes lvl assuming no positions are valid.

Finch.HollowData — Type

HollowData(lvl)

Represents a tensor which is represented by lvl but is sometimes entirely default(lvl).

Finch.LowerJulia — Method

TODO out of date dimensionalize!(prgm, ctx)

A program traversal which gathers dimensions of tensors based on shared indices. Index sharing is transitive, so A[i] = B[i] and B[j] = C[j] will induce a gathering of the dimensions of A, B, and C into one. The resulting dimensions are gathered into a Dimensions object, which can be accesed with an index name or a (tensor_name, mode_name) tuple.

The program is assumed to be in SSA form.

See also: virtual_size, virtual_resize, combinedim, TransformSSA

Finch.OpenScope — Type

OpenScope(ctx)

A transformation to initialize tensors that have just entered into scope.

See also: initialize!

Finch.RepeatData — Type

RepeatData(eltype, default)

Represents an array A[i] with many repeated runs of elements of type eltype and default default.

Finch.SolidData — Type

SolidData(lvl)

Represents a tensor which is represented by lvl

Finch.SparseData — Type

SparseData(lvl)

Represents a tensor A where A[i, :, ..., :] is sometimes entirely default(lvl) and is sometimes represented by lvl.

Finch.SubFiber — Type

SubFiber(lvl, pos)

SubFiber represents a fiber at position pos within lvl.

Finch.TransformSSA — Type

TransformSSA(freshen)

A transformation of a program to SSA form. Fresh names will be generated with freshen(name).

Finch.assemble_level! — Function

assemble_level!(lvl, ctx, pos, new_pos)

Assemble and positions pos+1:new_pos in lvl, assuming positions 1:pos were previously assembled.

Finch.base_rules — Method

base_rules(alg, ctx)

The basic rule set for Finch, uses the algebra to check properties of functions like associativity, commutativity, etc. Also assumes the context has a static hash names shash. This rule set simplifies, normalizes, and propagates constants, and is the basis for how Finch understands sparsity.

Finch.combinedim — Method

combinedim(ctx, a, b)

Combine the two dimensions a and b. To avoid ambiguity, only define one of

combinedim(ctx, ::A, ::B)
combinedim(ctx, ::B, ::A)

Finch.data_rep — Method

data_rep(tns)

Return a trait object representing everything that can be learned about the data based on the storage format (type) of the tensor

Finch.default — Function

default(fbr)

The default for a fiber is the value that each element of the fiber will have after initialization. This value is most often zero, and defaults to nothing.

See also: initialize!

Finch.f_code — Method

f_code(d) = DenseLevel.

Finch.f_code — Method

f_code(e) = ElementLevel.

Finch.f_code — Method

f_code(p) = PatternLevel.

Finch.f_code — Method

f_code(rl) = RepeatRLELevel.

Finch.f_code — Method

f_code(sc) = SparseCooLevel.

Finch.f_code — Method

f_code(sh) = SparseHashLevel.

Finch.f_code — Method

f_code(l) = SparseListLevel.

Finch.f_code — Method

f_code(sm) = SparseBytemapLevel.

Finch.f_code — Method

f_code(sv) = SparseVBLLevel.

Finch.ffindnz — Method

ffindnz(arr)

Return the nonzero elements of arr, as Finch understands arr. Returns (I, V), where I is a tuple of coordinate vectors, one for each mode of arr, and V is a vector of corresponding nonzero values, which can be passed to fsparse.

See also: (findnz)(https://docs.julialang.org/en/v1/stdlib/SparseArrays/#SparseArrays.findnz)

Finch.fiber — Function

fiber(arr, default = zero(eltype(arr)))

Copies an array-like object arr into a corresponding, similar Fiber datastructure. default is the default value to use for initialization and sparse compression.

See also: fiber!

Examples

julia> println(summary(fiber(sparse([1 0; 0 1]))))
2×2 @fiber(d(sl(e(0))))

julia> println(summary(fiber(ones(3, 2, 4))))
3×2×4 @fiber(d(d(d(e(0.0)))))

Finch.fiber! — Function

fiber!(arr, default = zero(eltype(arr)))

Like fiber, copies an array-like object arr into a corresponding, similar Fiber datastructure. However, fiber! reuses memory whenever possible, meaning arr may be rendered unusable.

Finch.fiber_ctr — Method

fiber_ctr(tns)

Return an expression that would construct a fiber suitable to hold data with a representation described by tns

Finch.freeze! — Method

freeze!(tns, ctx)

Freeze the update-only virtual tensor tns in the context ctx and return it. After freezing, the tensor is read-only.

Finch.freeze_level! — Method

freeze_level!(lvl, ctx, pos)

Freeze all fibers in lvl. Positions 1:pos need freezing.

Finch.fsparse — Function

fsparse(I::Tuple, V,[ M::Tuple, combine])

Create a sparse COO fiber S such that size(S) == M and S[(i[q] for i = I)...] = V[q]. The combine function is used to combine duplicates. If M is not specified, it is set to map(maximum, I). If the combine function is not supplied, combine defaults to + unless the elements of V are Booleans in which case combine defaults to |. All elements of I must satisfy 1 <= I[n][q] <= M[n]. Numerical zeros are retained as structural nonzeros; to drop numerical zeros, use dropzeros!.

See also: sparse

Examples

julia> I = ( [1, 2, 3], [1, 2, 3], [1, 2, 3]);

julia> V = [1.0; 2.0; 3.0];

julia> fsparse(I, V) SparseCoo (0.0) [1:3×1:3×1:3] │ │ │ └─└─└─[1, 1, 1] [2, 2, 2] [3, 3, 3] 1.0 2.0 3.0

Finch.fsparse! — Function

fsparse!(I::Tuple, V,[ M::Tuple])

Like fsparse, but the coordinates must be sorted and unique, and memory is reused.

Finch.fsprand — Method

fsprand([rng],[type], m::Tuple,p::AbstractFloat,[rfn])

Create a random sparse tensor of size m in COO format, in which the probability of any element being nonzero is independently given by p (and hence the mean density of nonzeros is also exactly p). Nonzero values are sampled from the distribution specified by rfn and have the type type. The uniform distribution is used in case rfn is not specified. The optional rng argument specifies a random number generator.

See also: (sprand)(https://docs.julialang.org/en/v1/stdlib/SparseArrays/#SparseArrays.sprand)

Examples

julia> fsprand(Bool, (3, 3), 0.5)
SparseCoo (false) [1:3,1:3]
├─├─[1, 1]: true
├─├─[3, 1]: true
├─├─[2, 2]: true
├─├─[3, 2]: true
├─├─[3, 3]: true  

julia> fsprand(Float64, (2, 2, 2), 0.5)
SparseCoo (0.0) [1:2,1:2,1:2]
├─├─├─[2, 2, 1]: 0.6478553157718558
├─├─├─[1, 1, 2]: 0.996665291437684
├─├─├─[2, 1, 2]: 0.7491940599574348 

Finch.fspzeros — Method

fspzeros([type], shape::Tuple)

Create a random zero tensor of size m, with elements of type type. The tensor is in COO format.

See also: (spzeros)(https://docs.julialang.org/en/v1/stdlib/SparseArrays/#SparseArrays.spzeros)

Examples

julia> fspzeros(Bool, (3, 3))
SparseCoo (false) [1:3,1:3]
    
julia> fspzeros(Float64, (2, 2, 2))
SparseCoo (0.0) [1:2,1:2,1:2]

Finch.get_reader — Method

get_reader(tns, ctx, protos...)

Return an object (usually a looplet nest) capable of reading the read-only virtual tensor tns. As soon as a read-only tensor enters scope, each subsequent read access will be initialized with a separate call to get_reader. protos is the list of protocols in each case.

Finch.get_updater — Method

get_updater(tns, ctx, protos...)

Return an object (usually a looplet nest) capable of updating the update-only virtual tensor tns. As soon as an update only tensor enters scope, each subsequent update access will be initialized with a separate call to get_updater. protos is the list of protocols in each case.

Finch.getindex_rep — Method

getindex_rep(tns, idxs...)

Return a trait object representing the result of calling getindex(tns, idxs...) on the tensor represented by tns.

Finch.getname — Function

getname(ex)

Return the name of the index expression ex. The name serves as a unique identifier and often corresponds to the variable name which holds a tensor. Tensors can have the same name only if they are === to each other. The names of indices are used to distinguish the loops they reference.

#TODO this function shouldn't exist

Finch.getresults — Function

getresults(stmt)

Return an iterator over the result tensors of an index expression. For example, where statements return the results of the consumer, not the producer, and assignments return their left hand sides.

Finch.getrules — Method

getrules(alg, ctx)

Return an array of rules to use for annihilation/simplification during compilation. One can dispatch on the alg trait to specialize the rule set for different algebras.

Finch.getsize — Function

virtual_size(tns, ctx)

Return a tuple of the dimensions of tns in the context ctx with access mode mode. This is a function similar in spirit to Base.axes.

Finch.getunbound — Method

getunbound(stmt)

Return an iterator over the names in an index expression that have yet to be bound.

julia> getunbound(@finch_program @loop i :a[i, j] += 2)
[j]
julia> getunbound(@finch_program i + j * 2 * i)
[i, j]

Finch.getvalue — Method

getvalue(ex)

If isliteral(ex) is true, return the value of ex. Defaults to the identity.

See also: isliteral

Finch.initialize! — Method

initialize!(tns, ctx)

Melt and initialize the read-only virtual tensor tns in the context ctx and return it. After melting, the tensor is update-only.

Finch.initialize_level! — Function

initialize_level!(lvl, ctx, pos)

Initialize and thaw all fibers within lvl, assuming positions 1:pos were previously assembled and frozen. The resulting level has no assembled positions.

Finch.isannihilator — Method

isannihilator(algebra, f, x)

Return true when f(a..., x, b...) = x in algebra.

Finch.isassociative — Method

isassociative(algebra, f)

Return true when f(a..., f(b...), c...) = f(a..., b..., c...) in algebra.

Finch.iscommutative — Method

iscommutative(algebra, f)

Return true when for all permutations p, f(a...) = f(a[p]...) in algebra.

Finch.isdistributive — Method

isidempotent(algebra, f)

Return true when f(a, b) = f(f(a, b), b) in algebra.

Finch.isidempotent — Method

isidempotent(algebra, f)

Return true when f(a, b) = f(f(a, b), b) in algebra.

Finch.isidentity — Method

isidentity(algebra, f, x)

Return true when f(a..., x, b...) = f(a..., b...) in algebra.

Finch.isinverse — Method

isinverse(algebra, f, g)

Return true when f(a, g(a)) is the identity under f in algebra.

Finch.isinvolution — Method

isinvolution(algebra, f)

Return true when f(f(a)) = a in algebra.

Finch.pattern! — Method

pattern!(fbr)

Return the pattern of fbr. That is, return a fiber which is true wherever fbr is structurally unequal to it's default. May reuse memory and render the original fiber unusable when modified.

julia> A = @fiber(sl(e(0.0), 10), [2.0, 0.0, 3.0, 0.0, 4.0, 0.0, 5.0, 0.0, 6.0, 0.0])
SparseList (0.0) [1:10]
├─[1]: 2.0
├─[3]: 3.0
├─[5]: 4.0
├─[7]: 5.0
├─[9]: 6.0

julia> pattern!(A)
SparseList (false) [1:10]
├─[1]: true
├─[3]: true
├─[5]: true
├─[7]: true
├─[9]: true

Finch.reassemble_level! — Function

reassemble_level!(lvl, ctx, pos_start, pos_end)

Set the previously assempled positions from pos_start to pos_end to level_default(lvl).

Finch.setname — Function

setname(ex, name)

Return a new expression, identical to ex, with the name name.

Finch.striplines — Method

striplines(ex)

Remove line numbers

Finch.trim! — Method

trim!(tns, ctx)

Before returning a tensor from the finch program, trim any excess overallocated memory.

Finch.unblock — Method

unblock(ex)

Flatten any redundant blocks into a single block, over the whole expression.

Finch.@fiber — Macro

@fiber ctr [arg]

Construct a fiber using abbreviated level constructor names. To override abbreviations, expressions may be interpolated with $. For example, Fiber(DenseLevel(SparseListLevel(Element(0.0)))) can also be constructed as @fiber(sl(d(e(0.0)))). Consult the documentation for the helper function f_code for a full listing of level format codes.

Optionally, an argument may be specified to copy into the fiber. This expression allocates. Use fiber(arg) for a zero-cost copy, if available.