Parallel Processing in Finch

Modelling the Architecture

Finch uses a simple, hierarchical representation of devices and tasks to model different kind of parallel processing. An AbstractDevice is a physical or virtual device on which we can execute tasks, which may each be represented by an AbstractTask.

Finch.AbstractTask — Type

AbstractTask

An individual processing unit on a device, responsible for running code.

Finch.AbstractDevice — Type

AbstractDevice

A datatype representing a device on which tasks can be executed.

The current task in a compilation context can be queried with get_task. Each device has a set of numbered child tasks, and each task has a parent task.

Finch.get_num_tasks — Function

get_num_tasks(dev::AbstractDevice)

Return the number of tasks on the device dev.

Finch.get_task_num — Function

get_task_num(task::AbstractTask)

Return the task number of task.

Finch.get_device — Function

get_device(task::AbstractTask)

Return the device that task is running on.

Finch.get_parent_task — Function

get_parent_task(task::AbstractTask)

Return the task which spawned task.

Data Transfer

Before entering a parallel loop, a tensor may reside on a single task, or represent a single view of data distributed across multiple tasks, or represent multiple separate tensors local to multiple tasks. A tensor's data must be resident in the current task to process operations on that tensor, such as loops over the indices, accesses to the tensor, or declare, freeze, or thaw. Upon entering a parallel loop, we must transfer the tensor to the tasks where it is needed. Upon exiting the parallel loop, we may need to combine the data from multiple tasks into a single tensor.

All tensor and buffer transfers are accomplished with the transfer function.

The distribute function is used by the compiler to orchestrate data distribution before and after a parallel region, with

different style objects signaling the type of transfer.

Note: After distributing a tensor, we must also update any in-progress traversals over the tensor that may appear throughout the program. This is done with the redistribute function. Tensors are responsible for defining their own redistribute behavior, but it should be guaranteed that distribute(tns, diff) == redistribute(tns, diff). In general, this means that any nested structure in the tensor should be preserved through transfers. Most subtensors will store a list of property names describing how to reach the subtensor from the root tensor.

Finch.distribute — Function

distribute(ctx, arr, device, diff, style)

If the virtual array is not on the given device, copy the array to that device. This function may modify underlying data arrays, but cannot change the virtual itself. This function is used to move data to the device before a kernel is launched. Since this function may modify the root node, iterators in-progress may need to be updated. We can store new root objects in the diff dictionary.

Finch.redistribute — Function

redistribute(ctx, node, diff)

When the root node is distributed, several iterators may need to be updated.

The redistribute function traverses tns and updates it based on the updated objects in the diff dictionary.

The distribute function is called on the Host and on the Device, and is responsible for distributing the tensor among tasks and collecting the results, if applicable.

If the tensor is a temporary tensor declared within the parallel loop, we distribute the tensor to Local scope. If the tensor is declared outside the parallel loop and is not modified, we distribute the tensor to Global scope. If the tensor is declared outside the parallel loop and is modified, we distribute the tensor to Shared scope. Depending on the architecture, several of these operations may be no-ops.

Finch.HostLocal — Type

HostLocal()

From the host, distribute the tensor to device local memory.

Finch.HostGlobal — Type

HostGlobal()

From the host, distribute the tensor to device global memory.

Finch.HostShared — Type

HostShared()

From the host, distribute the tensor to device shared memory.

Finch.DeviceLocal — Type

DeviceLocal()

From the device, load the local version of the tensor.

Finch.DeviceGlobal — Type

DeviceGlobal()

From the device, load the global view of the tensor.

Finch.DeviceShared — Type

DeviceShared()

From the device, load the shared view of the tensor.

The transfer function is used to distribute tensors and their constituent buffers to different memory spaces. We can ask for the default local, shared, or global memory spaces of an AbstractDevice with the local_memory, etc. trait functions.

Finch.transfer — Function

transfer(device, arr)

If the array is not on the given device, it creates a new version of this array on that device and copies the data in to it, according to the device trait. If the device is simply a data buffer, we copy the array into the buffer.

Finch.local_memory — Function

local_memory(dev::AbstractDevice)

Return the default local memory space of dev.

Finch.shared_memory — Function

shared_memory(dev::AbstractDevice)

Return the default shared memory space of dev.

Finch.global_memory — Function

global_memory(dev::AbstractDevice)

Return the default global memory space of dev.