Finch Compilation Pipeline

Finch Notation

Finch.FinchNotation.FinchNode — Type

FinchNode

A Finch IR node. Finch uses a variant of Concrete Index Notation as an intermediate representation.

The FinchNode struct represents many different Finch IR nodes. The nodes are differentiated by a FinchNotation.FinchNodeKind enum.

Finch.FinchNotation.literal — Constant

literal(val)

Finch AST expression for the literal value val.

Finch.FinchNotation.index — Constant

index(name)

Finch AST expression for an index named name. Each index must be quantified by a corresponding loop which iterates over all values of the index.

Finch.FinchNotation.variable — Constant

variable(name)

Finch AST expression for a variable named name. The variable can be looked up in the context.

Finch.FinchNotation.virtual — Constant

virtual(val)

Finch AST expression for an object val which has special meaning to the compiler. This type allows users to substitute their own ASTs, etc. into Finch expressions.

Finch.FinchNotation.value — Constant

value(val, type)

Finch AST expression for host code val expected to evaluate to a value of type type.

Finch.FinchNotation.loop — Constant

loop(idx, ext, body)

Finch AST statement that runs body for each value of idx in ext. Tensors in body must have ranges that agree with ext.

Finch.FinchNotation.assign — Constant

assign(lhs, op, rhs)

Finch AST statement that updates the value of lhs to op(lhs, rhs). The tensors of lhs are returned. Overwriting is accomplished with the function right(lhs, rhs) = rhs.

Finch.FinchNotation.call — Constant

call(op, args...)

Finch AST expression for the result of calling the function op on args....

Finch.FinchNotation.cached — Constant

cached(val, ref)

Finch AST expression val, equivalent to the quoted expression ref

Finch.FinchNotation.reader — Constant

reader()

Finch AST expression for an access mode that is read-only.

Finch.FinchNotation.updater — Constant

updater(mode)

Finch AST expression for an access mode that may modify the tensor. The mode field specifies whether this access returns this tensor (initializing and finalizing) or modifies it in place.

Finch.FinchNotation.create — Constant

create()

Finch AST expression for an "allocating" update. This access will initialize and freeze the tensor, and we can be sure that any values the tensor held before have been forgotten.

Finch.FinchNotation.modify — Constant

modify()

Finch AST expression for an "in place" update. The access will not initialize or freeze the tensor, but can modify it's existing values.

Finch.FinchNotation.declare — Constant

declare(tns, init)

Finch AST statement that declares tns with an initial value init in the current scope.

Finch.FinchNotation.freeze — Constant

freeze(tns)

Finch AST statement that freezes tns in the current scope.

Finch.FinchNotation.thaw — Constant

thaw(tns)

Finch AST statement that thaws tns in the current scope.

Finch.FinchNotation.forget — Constant

forget(tns)

Finch AST statement that marks the end of the lifetime of tns, at least until it is declared again.

Finch.FinchNotation.access — Constant

access(tns, mode, idx...)

Finch AST expression representing the value of tensor tns at the indices idx.... The mode differentiates between reads or updates and whether the access is in-place.

Finch.FinchNotation.sequence — Constant

sequence(bodies...)

Finch AST statement that executes each of it's arguments in turn. If the body is not a sequence, replaces accesses to read-only tensors in the body with getreader and accesses to update-only tensors in the body with getupdater.

Finch.FinchNotation.protocol — Constant

protocol(idx, mode)

Finch AST expression marking an indexing expression idx with the protocol mode. These usually reside at the toplevel of an indexing expression to an access.

Finch.FinchNotation.sieve — Constant

sieve(cond, body)

Finch AST statement that only executes body if cond is true.

Finch.FinchNotation.finch_leaf — Function

finch_leaf(x)

Return a terminal finch node wrapper around x. A convenience function to determine whether x should be understood by default as a literal, value, or virtual.

Program Instances

Finch relies heavily on Julia's metaprogramming capabilities ( macros and generated functions in particular) to produce code. To review briefly, a macro allows us to inspect the syntax of it's arguments and generate replacement syntax. A generated function allows us to inspect the type of the function arguments and produce code for a function body.

In normal Finch usage, we might call Finch as follows:

julia> C = @fiber(sl(e(0)));

julia> A = @fiber(sl(e(0)), [0, 2, 0, 0, 3]);

julia> B = @fiber(d(e(0)), [11, 12, 13, 14, 15]);

julia> @finch (C .= 0; @loop i C[i] = A[i] * B[i]);

julia> C
SparseList (0) [1:5]
├─[2]: 24
├─[5]: 45

The @macroexpand macro allows us to see the result of applying a macro. Let's examine what happens when we use the @finch macro (we've stripped line numbers from the result to clean it up):

julia> (@macroexpand @finch (C .= 0; @loop i C[i] = A[i] * B[i])) |> Finch.striplines |> Finch.regensym
quote
    _res_1 = (Finch.execute)((Finch.FinchNotation.sequence_instance)((Finch.FinchNotation.declare_instance)((Finch.FinchNotation.variable_instance)(:C, (Finch.FinchNotation.finch_leaf_instance)(C)), literal_instance(0)), begin
                    let i = index_instance(i)
                        (Finch.FinchNotation.loop_instance)(i, index_instance(:), (Finch.FinchNotation.assign_instance)((Finch.FinchNotation.access_instance)((Finch.FinchNotation.variable_instance)(:C, (Finch.FinchNotation.finch_leaf_instance)(C)), (Finch.FinchNotation.updater_instance)((Finch.FinchNotation.create_instance)()), (Finch.FinchNotation.variable_instance)(:i, (Finch.FinchNotation.finch_leaf_instance)(i))), (Finch.FinchNotation.literal_instance)((Finch.FinchNotation.initwrite)((Finch.default)(C))), (Finch.FinchNotation.call_instance)((Finch.FinchNotation.variable_instance)(:*, (Finch.FinchNotation.finch_leaf_instance)(*)), (Finch.FinchNotation.access_instance)((Finch.FinchNotation.variable_instance)(:A, (Finch.FinchNotation.finch_leaf_instance)(A)), (Finch.FinchNotation.reader_instance)(), (Finch.FinchNotation.variable_instance)(:i, (Finch.FinchNotation.finch_leaf_instance)(i))), (Finch.FinchNotation.access_instance)((Finch.FinchNotation.variable_instance)(:B, (Finch.FinchNotation.finch_leaf_instance)(B)), (Finch.FinchNotation.reader_instance)(), (Finch.FinchNotation.variable_instance)(:i, (Finch.FinchNotation.finch_leaf_instance)(i))))))
                    end
                end))
    begin
        C = Finch.get(_res_1, :C, C)
    end
    begin
        _res_1
    end
end

In the above output, @finch creates an AST of program instances, then calls Finch.execute on it. A program instance is a struct that contains the program to be executed along with its arguments. Although we can use the above constructors (e.g. loop_instance) to make our own program instance, it is most convenient to use the unexported macro Finch.finch_program_instance:

julia> using Finch: @finch_program_instance

julia> prgm = Finch.@finch_program_instance (C .= 0; @loop i C[i] = A[i] * B[i])
sequence_instance(declare_instance(variable_instance(:C, C), literal_instance(0)), loop_instance(index_instance(i), index_instance(:), assign_instance(access_instance(variable_instance(:C, C), updater_instance(create_instance()), index_instance(i)), literal_instance(Finch.FinchNotation.InitWriter{0}()), call_instance(variable_instance(:*, *), access_instance(variable_instance(:A, A), reader_instance(), index_instance(i)), access_instance(variable_instance(:B, B), reader_instance(), index_instance(i))))))

As we can see, our program instance contains not only the AST to be executed, but also the data to execute the program with. The type of the program instance contains only the program portion; there may be many program instances with different inputs, but the same program type. We can run our program using Finch.execute, which returns a NamedTuple of outputs.

julia> typeof(prgm)
Finch.FinchNotation.SequenceInstance{Tuple{Finch.FinchNotation.DeclareInstance{Finch.FinchNotation.VariableInstance{:C, Fiber{SparseListLevel{Int64, Int64, ElementLevel{0, Int64}}}}, Finch.FinchNotation.LiteralInstance{0}}, Finch.FinchNotation.LoopInstance{Finch.FinchNotation.IndexInstance{:i}, Finch.FinchNotation.IndexInstance{:(:)}, Finch.FinchNotation.AssignInstance{Finch.FinchNotation.AccessInstance{Finch.FinchNotation.VariableInstance{:C, Fiber{SparseListLevel{Int64, Int64, ElementLevel{0, Int64}}}}, Finch.FinchNotation.UpdaterInstance{Finch.FinchNotation.CreateInstance}, Tuple{Finch.FinchNotation.IndexInstance{:i}}}, Finch.FinchNotation.LiteralInstance{Finch.FinchNotation.InitWriter{0}()}, Finch.FinchNotation.CallInstance{Finch.FinchNotation.VariableInstance{:*, Finch.FinchNotation.LiteralInstance{*}}, Tuple{Finch.FinchNotation.AccessInstance{Finch.FinchNotation.VariableInstance{:A, Fiber{SparseListLevel{Int64, Int64, ElementLevel{0, Int64}}}}, Finch.FinchNotation.ReaderInstance, Tuple{Finch.FinchNotation.IndexInstance{:i}}}, Finch.FinchNotation.AccessInstance{Finch.FinchNotation.VariableInstance{:B, Fiber{DenseLevel{Int64, ElementLevel{0, Int64}}}}, Finch.FinchNotation.ReaderInstance, Tuple{Finch.FinchNotation.IndexInstance{:i}}}}}}}}}

julia> C = Finch.execute(prgm).C
SparseList (0) [1:5]
├─[2]: 24
├─[5]: 45

This functionality is sufficient for building finch kernels programatically. For example, if we wish to define a function pointwise_sum() that takes the pointwise sum of a variable number of vector inputs, we might implement it as follows:

julia> function pointwise_sum(As...)
           B = @fiber(d(e(0)))
           isempty(As) && return B
           i = Finch.FinchNotation.index_instance(:i)
           A_vars = [Finch.FinchNotation.variable_instance(Symbol(:A, n), As[n]) for n in 1:length(As)]
           #create a list of variable instances with different names to hold the input tensors
           ex = @finch_program_instance 0
           for A_var in A_vars
               ex = @finch_program_instance $A_var[i] + $ex
           end
           prgm = @finch_program_instance (B .= 0; @loop i B[i] = $ex)
           return Finch.execute(prgm).B
       end
pointwise_sum (generic function with 1 method)

julia> pointwise_sum([1, 2], [3, 4])
Dense [1:2]
├─[1]: 4
├─[2]: 6

Virtualization

TODO more on the way...

Tensor Life Cycle

Every virtual tensor must be in one of two modes: read-only mode or update-only mode. The following functions may be called on virtual tensors throughout their life cycle.

Finch.declare! — Function

declare!(tns, ctx, init)

Declare the read-only virtual tensor tns in the context ctx with a starting value of init and return it. Afterwards the tensor is update-only.

Finch.get_reader — Function

get_reader(tns, ctx, protos...)

Return an object (usually a looplet nest) capable of reading the read-only virtual tensor tns. As soon as a read-only tensor enters scope, each subsequent read access will be initialized with a separate call to get_reader. protos is the list of protocols in each case.

Finch.get_updater — Function

get_updater(tns, ctx, protos...)

Return an object (usually a looplet nest) capable of updating the update-only virtual tensor tns. As soon as an update only tensor enters scope, each subsequent update access will be initialized with a separate call to get_updater. protos is the list of protocols in each case.

Finch.freeze! — Function

freeze!(tns, ctx)

Freeze the update-only virtual tensor tns in the context ctx and return it. Afterwards, the tensor is read-only.

Finch.trim! — Function

trim!(tns, ctx)

Before returning a tensor from the finch program, trim any excess overallocated memory.

Finch.thaw! — Function

thaw!(tns, ctx)

Thaw the read-only virtual tensor tns in the context ctx and return it. Afterwards, the tensor is update-only.

Fiber internals

Fiber levels are implemented using the following methods:

Finch.default — Function

default(arr)

Return the initializer for arr. For SparseArrays, this is 0. Often, the default value becomes the fill or background value of a tensor.

Finch.declare_level! — Function

declare_level!(lvl, ctx, pos, init)

Initialize and thaw all fibers within lvl, assuming positions 1:pos were previously assembled and frozen. The resulting level has no assembled positions.

Finch.assemble_level! — Function

assemble_level!(lvl, ctx, pos, new_pos)

Assemble and positions pos+1:new_pos in lvl, assuming positions 1:pos were previously assembled.

Finch.reassemble_level! — Function

reassemble_level!(lvl, ctx, pos_start, pos_end)

Set the previously assempled positions from pos_start to pos_end to level_default(lvl).

Finch.freeze_level! — Function

freeze_level!(lvl, ctx, pos)

Freeze all fibers in lvl. Positions 1:pos need freezing.

Finch.level_ndims — Function

level_ndims(::Type{Lvl})

The result of level_ndims(Lvl) defines ndims for all subfibers in a level of type Lvl.

Finch.level_size — Function

level_size(lvl)

The result of level_size(lvl) defines the size of all subfibers in the level lvl.

Finch.level_axes — Function

level_axes(lvl)

The result of level_axes(lvl) defines the axes of all subfibers in the level lvl.

Finch.level_eltype — Function

level_eltype(::Type{Lvl})

The result of level_eltype(Lvl) defines eltype for all subfibers in a level of type Lvl.

Finch.level_default — Function

level_default(::Type{Lvl})

The result of level_default(Lvl) defines default for all subfibers in a level of type Lvl.