Public Functions

Finch.Environment — Type

Environment([parent]; kwargs...)

An environment can be thought of as the argument to a level that yeilds a fiber. Environments also allow parents levels to pass attributes to their children.

Finch.Fiber — Type

Fiber(lvl, env=Environment())

A fiber is a combination of a (possibly nested) level lvl and an environment env. The environment is often used to refer to a particular fiber within the level. Fibers are arrays, of sorts. The function refindex(fbr, i...) is used as a reference implementation of getindex for the fiber. Accessing an N-dimensional fiber with less than N indices will return another fiber.

Finch.Finalize — Type

Finalize(ctx)

A transformation to finalize output tensors before they leave scope and are returned to the caller.

See also: finalize!

Finch.Initialize — Type

Initialize(ctx)

A transformation to initialize tensors that have just entered into scope.

See also: initialize!

Finch.LowerJulia — Method

TODO out of date dimensionalize!(prgm, ctx)

A program traversal which gathers dimensions of tensors based on shared indices. Index sharing is transitive, so A[i] = B[i] and B[j] = C[j] will induce a gathering of the dimensions of A, B, and C into one. The resulting dimensions are gathered into a Dimensions object, which can be accesed with an index name or a (tensor_name, mode_name) tuple.

The program is assumed to be in SSA form.

See also: getsize, getsites, combinedim, TransformSSA

Finch.TransformSSA — Type

TransformSSA(freshen)

A transformation of a program to SSA form. Fresh names will be generated with freshen(name).

Finch.VirtualEnvironment — Type

VirtualEnvironment([parent]; kwargs...)

In addition to holding information about the environment instance itself, virtual environments may also hold information about the scope that this fiber lives in.

Finch.VirtualFiber — Type

VirtualFiber(lvl, env)

A virtual fiber is the avatar of a fiber for the purposes of compilation. Two fibers should share a name only if they hold the same data. lvl is a virtual object representing the level nest and env is a virtual object representing the environment.

Finch.combinedim — Method

combinedim(a, b)

Combine the two dimensions a and b. To avoid ambiguity, only define one of

combinedim(::A, ::B)
combinedim(::B, ::A)

Finch.default — Function

default(fbr)

The default for a fiber is the value that each element of the fiber will have after initialization. This value is most often zero, and defaults to nothing.

See also: initialize!

Finch.envcoordinate — Method

envcoordinate(env)

Get the coordinate (index) in the previous environment.

Finch.envdefaultcheck — Method

envdefaultcheck(env)

Return a variable which should be set to false if the subfiber is not entirely default.

Finch.envdepth — Method

envdepth()

Return the number of accesses (coordinates) unfurled so far in this environment.

Finch.envexternal — Method

envexternal(env)

Strip environments which are internal to the level, leaving the parent environment of the level.

Finch.envname — Method

envname()

The name of the tensor when it was last named.

Finch.envposition — Method

envposition(env)

Get the position in the environment. The position is an integer identifying which fiber to access in a level.

Finch.envreinitialized — Method

envreinitialized(env)

did the previous level selectively initialize this one?

Finch.f_code — Method

f_code(d) = DenseLevel.

Finch.f_code — Method

f_code(e) = ElementLevel.

Finch.f_code — Method

f_code(p) = PatternLevel.

Finch.f_code — Method

f_code(rl) = RepeatRLELevel.

Finch.f_code — Method

f_code(sc) = SparseCooLevel.

Finch.f_code — Method

f_code(sh) = SparseHashLevel.

Finch.f_code — Method

f_code(l) = SparseListLevel.

Finch.f_code — Method

f_code(sm) = SparseBytemapLevel.

Finch.f_code — Method

f_code(sv) = SparseVBLLevel.

Finch.ffindnz — Method

ffindnz(arr)

Return the nonzero elements of arr, as Finch understands arr. Returns (I, V), where I is a tuple of coordinate vectors, one for each mode of arr, and V is a vector of corresponding nonzero values, which can be passed to fsparse.

See also: (findnz)(https://docs.julialang.org/en/v1/stdlib/SparseArrays/#SparseArrays.findnz)

Finch.fiber — Function

fiber(arr, default = zero(eltype(arr)))

Copies an array-like object arr into a corresponding, similar Fiber datastructure. default is the default value to use for initialization and sparse compression.

See also: fiber!

Examples

julia> println(summary(fiber(sparse([1 0; 0 1]))))
2×2 @fiber(d(sl(e(0))))

julia> println(summary(fiber(ones(3, 2, 4))))
3×2×4 @fiber(d(d(d(e(0.0)))))

Finch.fiber! — Function

fiber!(arr, default = zero(eltype(arr)))

Like fiber, copies an array-like object arr into a corresponding, similar Fiber datastructure. However, fiber! reuses memory whenever possible, meaning arr may be rendered unusable.

Finch.finalize! — Method

finalize!(fbr, ctx, mode, idxs...)

Finalize the virtual fiber in the context ctx with access mode mode. Return the new fiber object.

Finch.finalize_level! — Function

finalize_level!(fbr, ctx, mode)

Finalize the level within the virtual fiber. These are the bulk cleanup steps.

Finch.fsparse — Function

fsparse(I::Tuple, V,[ M::Tuple, combine])

Create a sparse COO fiber S such that size(S) == M and S[(i[q] for i = I)...] = V[q]. The combine function is used to combine duplicates. If M is not specified, it is set to map(maximum, I). If the combine function is not supplied, combine defaults to + unless the elements of V are Booleans in which case combine defaults to |. All elements of I must satisfy 1 <= I[n][q] <= M[n]. Numerical zeros are retained as structural nonzeros; to drop numerical zeros, use dropzeros!.

See also: sparse

Examples

julia> I = ( [1, 2, 3], [1, 2, 3], [1, 2, 3]);

julia> V = [1.0; 2.0; 3.0];

julia> fsparse(I, V) SparseCoo (0.0) [1:3×1:3×1:3] │ │ │ └─└─└─[1, 1, 1] [2, 2, 2] [3, 3, 3] 1.0 2.0 3.0

Finch.fsparse! — Function

fsparse!(I::Tuple, V,[ M::Tuple])

Like fsparse, but the coordinates must be sorted and unique, and memory is reused.

Finch.fsprand — Method

fsprand([rng],[type], m::Tuple,p::AbstractFloat,[rfn])

Create a random sparse tensor of size m in COO format, in which the probability of any element being nonzero is independently given by p (and hence the mean density of nonzeros is also exactly p). Nonzero values are sampled from the distribution specified by rfn and have the type type. The uniform distribution is used in case rfn is not specified. The optional rng argument specifies a random number generator.

See also: (sprand)(https://docs.julialang.org/en/v1/stdlib/SparseArrays/#SparseArrays.sprand)

Examples

julia> fsprand(Bool, (3, 3), 0.5)
SparseCoo (false) [1:3×1:3]
│ │
└─└─[1, 1] [1, 3] [2, 2] [2, 3] [3, 3]
    true   true   true   true   true  

julia> fsprand(Float64, (2, 2, 2), 0.5)
SparseCoo (0.0) [1:2×1:2×1:2]
│ │ │
└─└─└─[1, 2, 2] [2, 1, 1] [2, 1, 2]
      0.647855  0.996665  0.749194 

Finch.fspzeros — Method

fspzeros([type], shape::Tuple)

Create a random zero tensor of size m, with elements of type type. The tensor is in COO format.

See also: (spzeros)(https://docs.julialang.org/en/v1/stdlib/SparseArrays/#SparseArrays.spzeros)

Examples

julia> fspzeros(Bool, (3, 3))
SparseCoo (false) [1:3×1:3]
│ │
└─└─
    
julia> fspzeros(Float64, (2, 2, 2))
SparseCoo (0.0) [1:2×1:2×1:2]
│ │ │
└─└─└─

Finch.getdefaultcheck — Method

getdefaultcheck(env)

Return a variable which should be set to false if the subfiber is not entirely default.

Finch.getname — Function

getname(ex)

Return the name of the index expression ex. The name serves as a unique identifier and often corresponds to the variable name which holds a tensor. Tensors can have the same name only if they are === to each other. The names of indices are used to distinguish the loops they reference.

Finch.getresults — Function

getresults(stmt)

Return an iterator over the result tensors of an index expression. For example, where statements return the results of the consumer, not the producer, and assignments return their left hand sides.

Finch.getsites — Function

getsites(tns)

Return an iterable over the identities of the modes of tns. If tns_2 is a transpose of tns, then getsites(tns_2) should be a permutation of getsites(tns) corresponding to the order in which modes have been permuted.

Finch.getsize — Function

getsize(tns, ctx, mode)

Return an iterable over the dimensions of tns in the context ctx with access mode mode. This is a function similar in spirit to Base.axes.

Finch.getunbound — Method

getunbound(stmt)

Return an iterator over the names in an index expression that have yet to be bound.

julia> getunbound(@finch_program @loop i :a[i, j] += 2)
[j]
julia> getunbound(@finch_program i + j * 2 * i)
[i, j]

Finch.getvalue — Method

getvalue(ex)

If isliteral(ex) is true, return the value of ex. Defaults to the identity.

See also: isliteral

Finch.hasdefaultcheck — Method

hasdefaultcheck(lvl)

Can the level check whether it is entirely default?

Finch.initialize! — Method

initialize!(fbr, ctx, mode)

Initialize the virtual fiber to it's default value in the context ctx with access mode mode. Return the new fiber object.

Finch.initialize_level! — Function

initialize_level!(fbr, ctx, mode)

Initialize the level within the virtual fiber to it's default value in the context ctx with access mode mode. Return the new level.

Finch.interval_assembly_depth — Method

interval_assembly_depth(lvl)

to what depth will the level tolerate interval environment properties for assembly?

Finch.pattern! — Method

pattern!(fbr)

Return the pattern of fbr. That is, return a fiber which is true wherever fbr is structurally unequal to it's default. May reuse memory and render the original fiber unusable when modified.

julia> A = Finch.Fiber(SparseList(10, [1, 6], [1, 3, 5, 7, 9], Element{0.0}([2.0, 3.0, 4.0, 5.0, 6.0])))
SparseList (0.0) [1:10]
│ 
└─[1] [3] [5] [7] [9]
  2.0 3.0 4.0 5.0 6.0

julia> pattern!(A)
SparseList (false) [1:10]
│ 
└─[1]  [3]  [5]  [7]  [9] 
  true true true true true

Finch.reinitializeable — Method

reinitializeable(lvl)

Does the level support selective initialization through assembly?

Finch.setname — Function

setname(ex, name)

Return a new expression, identical to ex, with the name name.

Finch.@fiber — Macro

@fiber ctr

Construct a fiber using abbreviated level constructor names. To override abbreviations, expressions may be interpolated with $. For example, Fiber(DenseLevel(SparseListLevel(Element(0.0)))) can also be constructed as @fiber(sl(d(e(0.0)))). Consult the documentation for the helper function f_code for a full listing of level format codes.